src/main/java1/ozamkovyi/web/servlet/adminServlets/AdminExchangeRateServlet.java
package ozamkovyi.web.servlet.adminServlets;

import ozamkovyi.db.dao.CurrencyDao;
import ozamkovyi.db.entity.Currency;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.servlet.http.HttpSession;
import java.io.IOException;
import java.util.ArrayList;

public class AdminExchangeRateServlet extends HttpServlet {
    @Override
    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        HttpSession session = req.getSession();
        Object page = session.getAttribute("pageNumber");
        int pageNumber = 0;
        int sortType = 0;
        if (page == null) {
            session.setAttribute("pageNumber", 1);
            session.setAttribute("sortType", 1);
            sortType = 1;
            pageNumber = 1;
        } else {
            pageNumber = (int) page;
            sortType = (int) session.getAttribute("sortType");
        }
        session.setAttribute("countCurrency", new CurrencyDao().getCountCurrency());
        session.setAttribute("listOfCurrency", new CurrencyDao().getAllCurrencyForChange(pageNumber, sortType));
        getServletContext().getRequestDispatcher("/jsp/adminExchangeRate.jsp").forward(req, resp);
    }

    @Override
    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        HttpSession session = req.getSession();
        if (req.getParameter("nextPage") != null) {
            Object pageNumber = session.getAttribute("pageNumber");
            if (pageNumber != null) {
                session.setAttribute("pageNumber", (int) pageNumber + 1);
            }
            resp.sendRedirect("/adminExchangeRate");
        }
        if (req.getParameter("previousPage") != null) {
            Object pageNumber = session.getAttribute("pageNumber");
            if (pageNumber != null) {
                session.setAttribute("pageNumber", (int) pageNumber - 1);
            }
            resp.sendRedirect("/adminExchangeRate");
        }

        if (req.getParameter("sortByName") != null) {
            Object sortType = session.getAttribute("sortType");
            if (sortType != null) {
                int sort = (int) sortType;
                if (sort == 1) {
                    session.setAttribute("sortType", 2);
                } else {
                    session.setAttribute("sortType", 1);
                }
            } else {
                session.setAttribute("sortType", null);
            }
            session.setAttribute("pageNumber", 1);
            resp.sendRedirect("/adminExchangeRate");
        }
        if (req.getParameter("sortByRate") != null) {
            Object sortType = session.getAttribute("sortType");
            if (sortType != null) {
                int sort = (int) sortType;
                if (sort == 3) {
                    session.setAttribute("sortType", 4);
                } else {
                    session.setAttribute("sortType", 3);
                }
            } else {
                session.setAttribute("sortType", null);
            }
            session.setAttribute("pageNumber", 1);
            resp.sendRedirect("/adminExchangeRate");
        }


        ArrayList<Currency> listOfCurrency = (ArrayList<Currency>) session.getAttribute("listOfCurrency");
        for (Currency currency : listOfCurrency) {
            if (req.getParameter("change " + currency.getId()) != null) {
                Object amountOdj = req.getParameter("amount " + currency.getId());
                double amount = 0;
                if (amountOdj != null) {
                    System.out.println("null");
                    amount = Double.parseDouble((String) amountOdj);
                }
                System.out.println(amount);
                new CurrencyDao().changeRate(currency.getId(), (float) amount);
                resp.sendRedirect("/adminExchangeRate");
            }
        }


        if (req.getParameter("buttonUnlockRequests") != null) {
            session.setAttribute("pageNumber", null);
            session.setAttribute("sortType", null);
            resp.sendRedirect("/adminHomepage");
        }
        if (req.getParameter("buttonAllUsers") != null) {
            session.setAttribute("pageNumber", null);
            session.setAttribute("sortType", null);
            resp.sendRedirect("/adminAllUsers");
        }
    }
}
